Split display routines from cli
All routines that decode and show the data are moved to a common file
separate from the command line interface.
