+++ Makefile
@@ -30,8 +30,8 @@ override CFLAGS += -DNVME_VERSION='"$(NVME_VERSION)"'
 
 NVME_DPKG_VERSION=1~`lsb_release -sc`
 
-nvme: nvme.c ./linux/nvme.h argconfig.o suffix.o common.o nvme-ioctl.o NVME-VERSION-FILE
-	$(CC) $(CPPFLAGS) $(CFLAGS) nvme.c $(LDFLAGS) -o $(NVME) argconfig.o suffix.o common.o nvme-ioctl.o
+nvme: nvme.c ./linux/nvme.h argconfig.o suffix.o nvme-print.o nvme-ioctl.o NVME-VERSION-FILE
+	$(CC) $(CPPFLAGS) $(CFLAGS) nvme.c $(LDFLAGS) -o $(NVME) argconfig.o suffix.o nvme-print.o nvme-ioctl.o
 
 nvme-ioctl.o: nvme-ioctl.c nvme-ioctl.h
 	$(CC) $(CPPFLAGS) $(CFLAGS) -c nvme-ioctl.c
@@ -42,7 +42,8 @@ argconfig.o: $(SRC)/argconfig.c $(SRC)/argconfig.h $(SRC)/suffix.h
 suffix.o: $(SRC)/suffix.c $(SRC)/suffix.h
 	$(CC) $(CPPFLAGS) $(CFLAGS) -c $(SRC)/suffix.c
 
-common.o: common.c
+nvme-print.o: nvme-print.c nvme-print.h
+	$(CC) $(CPPFLAGS) $(CFLAGS) -c nvme-print.c
 
 doc: $(NVME)
 	$(MAKE) -C Documentation
